refactor(server): isolate runtime state modules
Extract terminal logging and session bookkeeping behind dedicated modules so server.ts can focus on HTTP and WebSocket transport. Introduce the shared protocol definitions used by the new session registry.
